move puzzle study button to the board menu
