+++ ui/puzzle/src/view/boardMenu.ts
@@ -1,7 +1,9 @@
 import { h } from 'snabbdom';
 
 import { toggle } from 'lib';
-import { boardMenu as menuDropdown, boolPrefXhrToggle } from 'lib/view';
+import { renderNodesTxt } from 'lib/game/nodePGN';
+import * as licon from 'lib/licon';
+import { boardMenu as menuDropdown, boolPrefXhrToggle, hl } from 'lib/view';
 
 import type PuzzleCtrl from '../ctrl';
 
@@ -22,8 +24,43 @@ export default function (ctrl: PuzzleCtrl) {
       menu.voiceInput(boolPrefXhrToggle('voice', !!ctrl.voiceMove), true),
       menu.keyboardInput(boolPrefXhrToggle('keyboardMove', !!ctrl.keyboardMove), true),
     ]),
+    studyButton(ctrl),
     h('section.board-menu__links', [
       h('a', { attrs: { target: '_blank', href: '/account/preferences/display' } }, i18n.preferences.display),
     ]),
   ]);
 }
+
+const hiddenInput = (name: string, value: string) => h('input', { attrs: { type: 'hidden', name, value } });
+
+function renderPgnInput(ctrl: PuzzleCtrl): string {
+  const puzURL = `${location.origin}/training/${ctrl.data.puzzle.id}`;
+  const tags = [
+    ['Site', puzURL],
+    ['Variant', 'Standard'],
+    ['FEN', ctrl.initialNode.fen],
+  ]
+    .map(([k, v]) => `[${k} "${v}"]\n`)
+    .join('');
+  const linkPreamble = ` {${puzURL}} `;
+  return tags + linkPreamble + renderNodesTxt(ctrl.initialNode, true);
+}
+
+const studyButton = (ctrl: PuzzleCtrl) =>
+  ctrl.mode === 'play'
+    ? undefined
+    : h(
+        'section.board-menu__study',
+        hl('form', { attrs: { action: '/study/as', method: 'post', target: '_blank' } }, [
+          hiddenInput('pgn', renderPgnInput(ctrl)),
+          hiddenInput('fen', ctrl.initialNode.fen),
+          hiddenInput('orientation', ctrl.pov),
+          hiddenInput('variant', 'Standard'),
+          hiddenInput('mode', 'gamebook'),
+          hl(
+            'button.button.text',
+            { attrs: { type: 'submit', 'data-icon': licon.StudyBoard } },
+            i18n.site.toStudy,
+          ),
+        ]),
+      );
